add hometask4

+++ src/HomeWork3/Ability/MagicDamageSpells.java
@@ -3,5 +3,5 @@ package HomeWork3.Ability;
 import HomeWork3.Hero;
 
 public interface MagicDamageSpells {
-    int magicDamageSkill(Hero target);
+    double magicDamageSkill(Hero target);
 }
