+++ src/HomeWork3/Ability/MeleeAttack.java
@@ -3,5 +3,5 @@ package HomeWork3.Ability;
 import HomeWork3.Hero;
 
 public interface MeleeAttack {
-    int meleeSkill(Hero target);
+    double meleeSkill(Hero target);
 }
