+++ src/HomeWork3/Ability/ShootAttack.java
@@ -3,5 +3,5 @@ package HomeWork3.Ability;
 import HomeWork3.Hero;
 
 public interface ShootAttack {
-    int shootSkill(Hero target);
+    double shootSkill(Hero target);
 }
